Remove circular reference.
